Extract heading2, bnetcard, and registerguildcard
